:sparkles: Add support for saving a local copy of the bookmarks
This does nothing useful yet; this just ensures that when the bookmarks are
pulled down from the API, they're stored in a format that will work. The
next step with this will be to load this file on startup, display the
content, then go to the API and say "what's changed since?" and start to
refresh the local copy, updating things, adding things, removing things,
etc.

The aim being that there should be minimal traffic and, at a push, things
can work offline (the latter part coming later).