:sparkles: Swap to loading the local data if it exists
This is the next step. After this I want to tidy up the main flow. There's
now some duplication of effort going on here and that needs sorting out. But
this will be the plan:

If the local file exists
   Load and show that
   Find the latest modification in the data
   Ask the API for changes since then
   Merge them in
Else
   Get the whole list from the server
   Save it
End

At the moment this gives me the above, minus the "get changes since" part.
But... it's my holiday and I want dinner and nonsense on the TV.

Later!