:sparkles: Add the start of some filtering
Note that, as of right now, public and private don't work due to an oddity
in the XML data coming from pinboard.in, which in turn ends up making it
look like every bookmark is private, as seen by the library I'm building on
top of.

See https://github.com/bachya/aiopinboard/pull/265 for a proposed fix.