:hammer: Don't make the core message as non-bubble
It sort of made sense, but it really didn't given how I was using the
inherited messages; and this also shows that if you mark a message as
non-bubble, and the inherit from it, the child classes don't inherit the
bubble status.

Not sure if this is a bug in Textual or a deliberate design choice.