:sparkles: Add support for additive tag searching
This is the core of it, I'm not 100% sure about the interface just yet. The
idea here is you select a tag to just tag that tag, but you press `+` to add
a tag to the filter. This is fine on the keyboard but it's not a great mouse
experience.

I'll keep playing and see what I like; but the core code will work no
matter.