:construction: Revamping the Details pane
Mostly there with how I want it to look, but I need to spend a bit of time
thinking about how I'm going to do the tags. Using links in Markdown is
great and all, but it's a bit keyboard-hostile. I want to be
keyboard-friendly too.

So this is a WiP commit at the end of the evening, and the next time I'm in
here I'll have a proper think about how I'm going to do this.

See also https://github.com/Textualize/textual/issues/3687 as a related
issue that I've discovered while making these changes (it explains why the
scrollbar in the details pane isn't showing at the moment, if it's needed).