:construction: Initial work on a edit screen for bookmarks
Work-in-progress of adding an edit screen for bookmarks. This will be used
for adding new bookmarks and editing existing ones. Something that needs to
be added is a tag suggestion facility -- perhaps using the autocomplete or
something; need to look into that.

https://github.com/Textualize/textual/issues/3721 is currently causing the
layout to not be as good as it could be, and there's also the issue of how
TextArea works meaning that:

- There's no simple coloring.
- There's no easy way to handle going into light mode.
- There's no way to have the cursor not be visible when not focused.

Some of these things will just have to be so until changes are made in
Textual itself.