:sparkles: Start a system of command messages
Here I move the call to edit a bookmark into a message, and in doing so take
the binding for this off the main screen, and place bindings for posting
that message on the widgets where it makes sense; I then move the editing of
a bookmark into an event handler on the main screen.

The idea here is that the binding is only in place when a suitable widget
has focus, while at the same time keeping the logic for the edit in once
place.