:bug: Ensure completion of a tag doesn't nuke case of those before
The suggester API in Textual, if you're in case-insensitive mode, only ever
hands you the "normalised" value. Here I'm making it look like I'm doing
suggestions on the latest word, by completing on the whole input *after*
I've found a match for the last word in the input. This means that when I
was building the completion, the case of the already-entered tags would be
changed to the casefold version. Not what I wanted.

This fixes that by making the suggester case sensitive, and then doing all
the case-insensitive matching myself.