:lipstick: Tweak the sizing of the confirmation dialog
This is a bit of a hack, I can't quite get this to work how I want, but this
is close. Here's what I want:

- Dialog contains a Label.
- I want the Label to grow as wide as it needs, up to about 80% of the
  terminal width.
- I want the dialog itself to be as narrow as possible.
- If the Label hits the width limit I want it to wrap its content.

I've been in a fight with the max-width of the Vertical that makes the
dialog, and the Label, for a while now. I can't quite dial in the exact
style I need here. So, for now, I'm saying the Vertical can be at most 80%
of the terminal width, and the Label 70% of it. This is a kludge, but it
works in most cases and has the look I want.