:sparkles: Add the start of a proper context-sensitive help screen
This adds the idea of each screen and widget having a CONTEXT_HELP class
variable, and the help screen will look for that and make use of it to build
up the help text.

So far this just adds help to the main screen and the bookmarks widget, but
now I can add help to the different widgets in small neat changes.