:hammer: Switch to using a minimal internal client API
While aiopinboard is awesome, it's using the XML version of the API, which
has a couple of issues. Moreover, I'm thinking it'll be fun to get fine
control of my use of the API here. Also, I want to be as aggressive as
possible when it comes to *not* hitting in the API, and observing rate
limits, etc.

This initial re-working of Tinboard only partially takes rate limits into
account, but it gets me going on what I want.