:hammer: Simplify the search dialog
I think I'd started it out wrapped in a Vertical, assuming I was going to
add some other options (case sensitivity, that sort of thing). As it turns
out, I went with a "just go find everything I can" approach, and I prefer
it.

So this reduces the dialog to the bare minimum of widgets and code, for the
exact same interface.