:hammer: Update for Textual 0.47.0
In part this was intended to switch things up to using nested CSS where
possible, to see how well it works in the field. It did throw up one
possible bug:

   https://github.com/Textualize/textual/issues/3961

and I also can't get the token input dialog to use nested CSS without things
going crazy wrong; I need another run at that to try and isolate what's
going on and perhaps get it down to a MRE if it's a genuine bug.