:hammer: Make the from_json method of the bookmark data more generic
This should then handle data coming back from the API, and also any local
copy, work fine when being loaded up with from_json.

While this method is intended for and written for working with the Pinbaord
API itself, it's not impossible that the boolean values get turned into
bools somewhere along the way; so here I allow for that.

Accept shit, send diamonds, as the saying goes.