:hammer: Move bookmark data handling out of the Bookmarks widget
Initial work to implement #39. The idea here being that the bookmark data is
part of its own data structure now, rather than being a collection of
candidate prompts that are managed by the Bookmarks widget.

I feel there's a wee bit of cruft left over given this has been done "after
the fact", but this is a start of the cleanup.

This will also make it easier to make all the local tags available to the
inline addition feature.

I think a followup change to this might be to move the
loading/saving/downloading of bookmarks fully outwith the Bookmarks widget.