:sparkles: Add support for copying to a remote clipboard
Where possible anyway. The idea here is we're use the newly-added textual
copy_to_clipboard method to get the bookmark as far as it'll go, and then we
use pyperclip anyway to get it to the "local" clipboard, which will mean
that even on terminals that don't support OSC 52, it should get somewhere
useful anyway.